Merge latest

+++ app/src/browser/application.es6
@@ -1,6 +1,6 @@
 /* eslint global-require: "off" */
 
-import { BrowserWindow, Menu, app, ipcMain, dialog, powerMonitor } from 'electron';
+import { BrowserWindow, Menu, app, ipcMain, dialog } from 'electron';
 
 import fs from 'fs-plus';
 import url from 'url';
@@ -12,6 +12,7 @@ import WindowManager from './window-manager';
 import FileListCache from './file-list-cache';
 import ConfigMigrator from './config-migrator';
 import ApplicationMenu from './application-menu';
+import ApplicationTouchBar from './application-touch-bar';
 import AutoUpdateManager from './autoupdate-manager';
 import SystemTrayManager from './system-tray-manager';
 import DefaultClientHelper from '../default-client-helper';
@@ -100,6 +101,9 @@ export default class Application extends EventEmitter {
       initializeInBackground: initializeInBackground,
     });
     this.systemTrayManager = new SystemTrayManager(process.platform, this);
+    if (process.platform === 'darwin') {
+      this.touchBar = new ApplicationTouchBar(resourcePath);
+    }
 
     this.setupJavaScriptArguments();
     this.handleEvents();
@@ -114,7 +118,8 @@ export default class Application extends EventEmitter {
   }
 
   getMainWindow() {
-    return this.windowManager.get(WindowManager.MAIN_WINDOW).browserWindow;
+    const win = this.windowManager.get(WindowManager.MAIN_WINDOW);
+    return win ? win.browserWindow : null;
   }
 
   getAllWindowDimensions() {
@@ -164,6 +169,9 @@ export default class Application extends EventEmitter {
   }
 
   async oneTimeMoveToApplications() {
+    if (process.platform !== 'darwin') {
+      return;
+    }
     if (this.devMode || this.specMode) {
       return;
     }
@@ -240,61 +248,31 @@ export default class Application extends EventEmitter {
     }
   }
 
-  _relaunchToInitialWindows = ({ resetConfig, resetDatabase } = {}) => {
-    // This will re-fetch the NylasID to update the feed url
-    this.autoUpdateManager.updateFeedURL();
-    this.windowManager.destroyAllWindows();
+  _resetDatabaseAndRelaunch = ({ errorMessage } = {}) => {
+    if (this._resettingAndRelaunching) return;
+    this._resettingAndRelaunching = true;
 
-    let fn = callback => callback();
-    if (resetDatabase) {
-      fn = this._deleteDatabase;
+    if (errorMessage) {
+      dialog.showMessageBox({
+        type: 'warning',
+        buttons: ['Okay'],
+        message: `We encountered a problem with your local email database. We will now attempt to rebuild it.`,
+        detail: errorMessage,
+      });
     }
 
-    fn(async () => {
-      if (resetDatabase) {
-        // TODO BG Reset Database via helpers
-      }
-      if (resetConfig) {
-        this.config.set('nylas', null);
-        this.config.set('edgehill', null);
-      }
-      this.openWindowsForTokenState();
-    });
+    const done = () => {
+      app.relaunch();
+      app.quit();
+    };
+    this.windowManager.destroyAllWindows();
+    this._deleteDatabase(done);
   };
 
   _deleteDatabase = callback => {
     this.deleteFileWithRetry(path.join(this.configDirPath, 'edgehill.db'), callback);
     this.deleteFileWithRetry(path.join(this.configDirPath, 'edgehill.db-wal'));
     this.deleteFileWithRetry(path.join(this.configDirPath, 'edgehill.db-shm'));
-  };
-
-  rebuildDatabase = ({ showErrorDialog = true, detail = '' } = {}) => {
-    if (this._rebuildingDatabase) {
-      return;
-    }
-    this._rebuildingDatabase = true;
-    if (showErrorDialog) {
-      dialog.showMessageBox({
-        type: 'warning',
-        buttons: ['Okay'],
-        message: `We encountered a problem with your local email database. We will now attempt to rebuild it.`,
-        detail,
-      });
-    }
-
-    // We need to set a timeout so `rebuildDatabases` immediately returns.
-    // If we don't immediately return the main window caller wants to wait
-    // for this function to finish so it can get the return value via ipc.
-    // Unfortunately since this function destroys the main window
-    // immediately, an error will be thrown.
-    setTimeout(() => {
-      this.windowManager.destroyAllWindows();
-      this._deleteDatabase(async () => {
-        // TODO BG Invoke db helepr
-        this._rebuildingDatabase = false;
-        this.openWindowsForTokenState();
-      });
-    }, 0);
   };
 
   // Registers basic application commands, non-idempotent.
@@ -334,7 +312,7 @@ export default class Application extends EventEmitter {
       );
     });
 
-    this.on('application:relaunch-to-initial-windows', this._relaunchToInitialWindows);
+    this.on('application:reset-database', this._resetDatabaseAndRelaunch);
 
     this.on('application:quit', () => {
       app.quit();
@@ -442,10 +420,6 @@ export default class Application extends EventEmitter {
       });
     }
 
-    powerMonitor.on('resume', () => {
-      this.windowManager.sendToAllWindows('app-resumed-from-sleep', {});
-    });
-
     app.on('window-all-closed', () => {
       this.windowManager.quitWinLinuxIfNoWindows();
     });
@@ -549,6 +523,9 @@ export default class Application extends EventEmitter {
     ipcMain.on('update-application-menu', (event, template, keystrokesByCommand) => {
       const win = BrowserWindow.fromWebContents(event.sender);
       this.applicationMenu.update(win, template, keystrokesByCommand);
+      if (win === this.getMainWindow() && process.platform === 'darwin') {
+        this.touchBar.update(template);
+      }
     });
 
     ipcMain.on('command', (event, command, ...args) => {
